auth-service/requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2==2.9.3
python-dotenv==0.20.0